% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais_to_obis.R
\name{aisIndex_to_OBIS_occurenceCore}
\alias{aisIndex_to_OBIS_occurenceCore}
\title{AIS Biofouling to OBIS Occurence Darwin Core (DwC)}
\usage{
aisIndex_to_OBIS_occurenceCore(file)
}
\arguments{
\item{file}{CSV formatted file created using \code{ais_all_presence}}
}
\description{
AIS Biofouling to OBIS Occurence Darwin Core (DwC)
}
\examples{
aisIndex_to_OBIS_occurenceCore("file.csv")

}
