% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COERS_TDR_proc.R
\name{tdr_proc}
\alias{tdr_proc}
\title{SABS wharf long term temperature logging data transformation function}
\usage{
tdr_proc(x)
}
\arguments{
\item{x}{a comma delimited text file created from the Hobo temperature logger
manufacturer software with the following columns `#`, `Date Time,
GMT+00:00`, `Abs Pres, psi (LGR S/N: xxxx, SEN S/N: xxxx)`, `Temp, °C (LGR
S/N: xxxx, SEN S/N: xxxx)`, `Water Level, meters (LGR S/N: xxxx)`}
}
\value{
returns a comma delimited text file in the format
  `[YEAR]-[MONTH]-[DAY]_SABS_warf_water_temperature_data.csv` for importing
  into COERS.TDRLOG
}
\description{
function to transform Hobo temperature logger data that has been
  converted from proprietary binary format to a comma separated text file
  into a format that can be ingested in to the COERS.TDRLOG Oracle database
  tables on the DFO production database (PTRAN).
}
\examples{
tdr_proc('hobo_data_file.csv')
}
\author{
Peter Kraska, \email{peter.kraska@dfo-mpo.gc.ca}
}
