% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COERS_Drifter_Tracking.R
\name{drifter_tracking}
\alias{drifter_tracking}
\title{SPOT Drifter tracking script using the SPOT API}
\usage{
drifter_tracking(interval = 120, password)
}
\arguments{
\item{interval}{The interval to poll the SPOT website, in seconds. Default is
120 seconds as that should not be flagged by the service provider.}

\item{password}{The SPOT portal login password}
}
\value{
returns a comma delimited text file named drifter_coords.csv to the
  "C:/DFO-MPO/" folder on DFO computers.
}
\description{
SPOT Drifter tracking script using the SPOT API
}
\examples{
drifter_tracking(interval = 120, password = 'something')
}
