% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo2nc_convert.R
\name{hobo2nc_convert}
\alias{hobo2nc_convert}
\title{Hobo Temperature Logger to netCDF conversion}
\usage{
hobo2nc_convert(
  file,
  atm_corr_file = NA,
  latitude,
  longitude,
  deployment,
  destination = c("R:/Science/CESD/CESD_DataManagement/data_out/HOBO", "."),
  project_lead = "",
  project = "",
  cf_title = "Maritime Ecosystem and Ocean Science Hobo temperature logger data",
  cf_institute = c("BIO", "SABS"),
  cf_source =
    "Hobo temperature logger data collected as part of Fisheries and Oceans Canada (DFO) science activities in the Maritime region of Canada",
  cf_history = "",
  cf_references = "",
  cf_comment = "",
  cf_author = "Peter Kraska, CESD Divisional data Manager <Peter.Kraska@DFO-MPO.gc.ca>",
  cf_conventions = "CF-1.8",
  cf_calendar = "standard",
  creation_date = format(x = Sys.time(), tz = "GMT", format = "\%FT\%H:\%M\%SZ")
)
}
\arguments{
\item{file}{A character string path to a Hobo temperatuer and pressure logger
csv output file}

\item{atm_corr_file}{(optional) if an atmospheric pressure correction Hobo
temperature and pressure logger was deployed, list the character string
path to the file here}

\item{latitude}{(required to compute depth using the TEOS-10 algorithm,
https://www.teos-10.org/pubs/gsw/html/gsw_z_from_p.html) Latitude of the
Hobo Temperature and Presseure Logger}

\item{longitude}{longitude in WGS84}

\item{deployment}{Character string, how the temperature was deployed (e.g. "affixed to floating dock, 1m below water level.")}

\item{destination}{an existing folder to save netCDF files to}

\item{project_lead}{an optional field to add the project lead's name to the
file global attributes}

\item{project}{Character string indicating the funding source and or
overarching project this data was collected as part of}

\item{cf_title}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation). A
succinct description of what is in the dataset.}

\item{cf_institute}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation).
Specifies where the original data was produced.}

\item{cf_source}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation). The
method of production of the original data. If it was model-generated,
source should name the model and its version, as specifically as could be
useful. If it is observational, source should characterize it (e.g.,
"surface observation" or "radiosonde").}

\item{cf_history}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation).
Provides an audit trail for modifications to the original data.
Well-behaved generic netCDF filters will automatically append their name
and the parameters with which they were invoked to the global history
attribute of an input netCDF file. We recommend that each line begin with a
timestamp indicating the date and time of day that the program was
executed.}

\item{cf_references}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation).
Published or web-based references that describe the data or methods used to
produce it.}

\item{cf_comment}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation).
Miscellaneous information about the data or methods used to produce it.}

\item{cf_author}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation)}

\item{cf_conventions}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation)}

\item{cf_calendar}{an optional field to add to the generated NetCDF file's
global attributes (required for CF compliant NetCDF file creation).
Specifies the calendar used for time dimensions.
(http://cfconventions.org/Data/cf-conventions/cf-conventions-1.10/cf-conventions.html#calendar)}

\item{creation_date}{creation date to append to the global attributes of the
NetCDF file}
}
\value{
\code{hobo2nc_convert()} creates netCDF files from castaway CTD CSV
  files
}
\description{
Hobo Temperature Logger to netCDF conversion
}
\examples{
hobo2nc_convert("data/")

}
