% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lttm_convert_odf_to_nc.R
\name{lttm_convert_odf_to_nc}
\alias{lttm_convert_odf_to_nc}
\title{Convert Long Term Temperature Monitoring (LTTM) Ocean Data Format (ODF) to
NetCDF (.nc) formatted files}
\usage{
lttm_convert_odf_to_nc(input, output)
}
\arguments{
\item{input}{character vector path to ODF formatted Long Term Temperature
Monitoring (LTTM) file}

\item{output}{character vector path to directory to save NetCDF file in}
}
\description{
Convert Long Term Temperature Monitoring (LTTM) Ocean Data Format (ODF) to
NetCDF (.nc) formatted files
}
\examples{
lttm_convert_odf_to_nc("data/odf/file.odf")

}
