% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf2nc_convert.R
\name{odf2nc_convert}
\alias{odf2nc_convert}
\title{Title}
\usage{
odf2nc_convert(
  file,
  cf_title,
  cf_institute = c("BIO", "SABS"),
  cf_source =
    "[Insert information here] data collected as part of Fisheries and Oceans Canada (DFO) science activities in the Maritime region of Canada",
  cf_history = "",
  cf_references = "",
  cf_comment = "",
  cf_author = Sys.getenv("USERNAME"),
  cf_conventions = "CF-1.8",
  cf_calendar = "standard",
  creation_date = format(x = Sys.time(), tz = "GMT", format = "\%FT\%H:\%M\%SZ")
)
}
\arguments{
\item{file}{A character string path to a Hobo temperatuer and pressure logger
csv output file}

\item{creation_date}{}
}
\description{
Title
}
