% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COERS_wind_proc.R
\name{wind_proc}
\alias{wind_proc}
\title{SABS wharf long term wind speed and direction logging data transformation
function}
\usage{
wind_proc(
  x,
  station = 1,
  gaugenum = 999,
  latitude = 45.08275,
  longitude = 67.08508,
  area = "SABS Wharf"
)
}
\arguments{
\item{x}{comma delimited wind speed and direction file from the SABS wharf
wind speed and direction logger}

\item{station}{`STATION` to be used for identification in the
`COERS.WINDDATA` table on `PTRAN`. Default value is `1` (only value in the
table).}

\item{gaugenum}{`GAUGENUM` probably wind speed and direction logger
identification number, but only `999` present in the the `COERS.WINDDATA`
table}

\item{latitude}{Latitude of wind speed and direction logger unit measurements
in decimal degrees WGS84}

\item{longitude}{Longitude of wind speed and direction logger unit
measurements in decimal degrees WGS84}

\item{area}{Character string for location of wind speed and direction logger}
}
\value{
Returns a comma delimited text file in
  `R:/Science/CESD/COERS/FPage/data/Wind logger/FOR ORACLE IMPORT/` for
  upload to PTRAN by COERS staff.


}
\description{
SABS wharf long term wind speed and direction logging data transformation
function
}
\examples{
wind_proc(x)

}
