% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais_all_presence.R
\name{ais_all_presence}
\alias{ais_all_presence}
\title{Aquatic Invasive Species (AIS) annual biofouling presence/absence data
amalgamation}
\usage{
ais_all_presence(input, output)
}
\arguments{
\item{input}{character string to a directory holding only the final yearly
summary data for the presence/absence data in CSV format}

\item{output}{character string indicating the folder to save the output CSV
to. CSV file is called \code{AIS_biofouling_allyears_presence_absence.csv}
and function will create a folder from \code{output}.}
}
\description{
Aquatic Invasive Species (AIS) annual biofouling presence/absence data
amalgamation
}
