% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf2nc_convert.R
\name{odf2nc_convert}
\alias{odf2nc_convert}
\title{Title}
\usage{
odf2nc_convert(
  file,
  cf_title = "Maritime Ecosystem and Ocean Science Hobo temperature logger data",
  cf_institute = c("BIO", "SABS"),
  cf_source =
    "Hobo temperature logger data collected as part of Fisheries and Oceans Canada (DFO) science activities in the Maritime region of Canada",
  cf_history = "",
  cf_references = "",
  cf_comment = "",
  cf_author = Sys.getenv("USERNAME"),
  cf_conventions = "CF-1.8",
  cf_calendar = "standard",
  creation_date = format(x = Sys.time(), tz = "GMT", format = "\%FT\%H:\%M\%SZ")
)
}
\arguments{
\item{file}{}

\item{cf_title}{}

\item{cf_institute}{}

\item{cf_source}{}

\item{cf_history}{}

\item{cf_references}{}

\item{cf_comment}{}

\item{cf_author}{}

\item{cf_conventions}{}

\item{cf_calendar}{}

\item{creation_date}{}
}
\description{
Title
}
